refactor(pypi): split out code for env marker evaluation for reuse
This is just a small PR to reduce the scope of #2059.

Work towards #260, #1105, #1868.
