+++ python/private/pypi/whl_installer/BUILD.bazel
@@ -5,6 +5,7 @@ py_library(
     srcs = [
         "arguments.py",
         "namespace_pkgs.py",
+        "platform.py",
         "wheel.py",
         "wheel_installer.py",
     ],
